+++ src/lib/components/formBuilder/MainPollSettings.svelte
