+++ src/lib/components/formBuilder/PollSectionSettings.svelte
