+++ src/routes/new/+page.svelte
@@ -1,7 +1,7 @@
 <script lang="ts">
 	import { goto } from '$app/navigation';
-	import MainPollSettings from '$lib/components/form/MainPollSettings.svelte';
-	import PollSectionSettings from '$lib/components/form/PollSectionSettings.svelte';
+	import MainPollSettings from '$lib/components/formBuilder/MainPollSettings.svelte';
+	import PollSectionSettings from '$lib/components/formBuilder/PollSectionSettings.svelte';
 	import type { Poll } from '$lib/schemas/poll';
 	import {
 		createMainPollStore,
